fix(server): a live turn means deliver — the gate stops guessing worker disposition
Reverts the steer/cancel exemption and deletes the deferral it was patching.
Both were the same mistake in opposite directions.

The exemption asked `targetsActiveTurn(payload)` whether the worker would
consume a message as control. It cannot know. The worker additionally requires
`currentWorker`, `messageBatcher.isCurrentlyProcessing()`, and
`currentWorkerUserId === data.userId` — and a shared channel thread is served by
ONE worker, so a second participant's message fails that last check as a matter
of course and falls through to new work. The exemption therefore promised a
protection it could not keep.

But the deferral it replaced was also wrong, for the reason the exemption was
written: `currentWorker.steer()` splices a follow-up into the live turn and
`currentWorker.cancel()` aborts it, so withholding those messages degrades
steering into queueing and makes /cancel silently do nothing until the turn ends.
Fourteen review rounds missed that because every reviewer read the gateway and
none read the worker.

There is no payload-only answer, so the gate stops asking: a stale worker
mid-turn is delivered to, unrecycled. That leaves this one case at exactly its
pre-gate behaviour — no regression — and the recycle is not lost, it runs on the
first claim that finds the worker quiet. The cost, now stated in the module doc
instead of implied: a conversation that is never quiet at claim time is never
recycled. Closing that needs worker-confirmed delivery disposition, which is the
turn-lifecycle consolidation's job, not a heuristic here.

Net effect on the branch: `prior-turn-live` leaves the deferral union,
packages/core/src/turn-control.ts and its export are deleted, and the worker
keeps its own predicates (the move was only ever justified by the gate needing
to share them).
